-memory and SQUEAK_MEMORY handle more than 31 bits of extra memory on 64-bit architectures
git-svn-id: http://squeakvm.org/svn/squeak/trunk@2763 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: f89071c8856fec8e892bfb2290c8ec143ba88227

+++ platforms/unix/vm/sqUnixMain.c
@@ -27,7 +27,7 @@
 
 /* Author: Ian Piumarta <ian.piumarta@squeakland.org>
  *
- * Last edited: 2012-09-16 21:47:24 by piumarta on linux64
+ * Last edited: 2013-08-10 16:23:03 by piumarta on emilia
  */
 
 #include "sq.h"
@@ -84,7 +84,7 @@ static char **vmArgVec=		0;
 static int    squeakArgCnt=	0;
 static char **squeakArgVec=	0;
 
-static int    extraMemory=	0;
+static long   extraMemory=	0;
        int    useMmap=		DefaultMmapSize * 1024 * 1024;
 
 static int    useItimer=	1;	/* 0 to disable itimer-based clock */
@@ -940,10 +940,10 @@ static void loadModules(void)
 /* built-in main vm module */
 
 
-static int strtobkm(const char *str)
+static long strtobkm(const char *str)
 {
   char *suffix;
-  int value= strtol(str, &suffix, 10);
+  long value= strtol(str, &suffix, 10);
   switch (*suffix)
     {
     case 'k': case 'K':
@@ -1310,9 +1310,9 @@ void imgInit(void)
       else
 	extraMemory= DefaultHeapSize * 1024 *1024;
 #    ifdef DEBUG_IMAGE
-      printf("image size %d + heap size %d (useMmap = %d)\n", (int)sb.st_size, extraMemory, useMmap);
+      printf("image size %d + heap size %ld (useMmap = %d)\n", (int)sb.st_size, extraMemory, useMmap);
 #    endif
-      extraMemory += (int)sb.st_size;
+      extraMemory += (long)sb.st_size;
       readImageFromFileHeapSize(f, extraMemory);
       sqImageFileClose(f);
       break;
